Revert "Changed data type of Pressure and Temperature properties in MS5837Result to long"
This reverts commit f2dc6c812ffa3c47944990b8363fee46e64724ab.

+++ Source/Monoculture.TinyCLR.Drivers.MS5837/MS5837Result.cs
@@ -18,7 +18,7 @@ namespace Monoculture.TinyCLR.Drivers.MS5837
 {
     public class MS5837Result
     {
-        public long Pressure { get; internal set; }
-        public long Temperature { get; internal set; }
+        public double Pressure { get; internal set; }
+        public double Temperature { get; internal set; }
     }
 }
